fix inst count logic

+++ hook-api-examples/hookapi.h
@@ -29,7 +29,7 @@ extern int32_t _g                  (uint32_t id, uint32_t maxiter);
  * @param read_len The length of the string. May be 0.
  * @return Will never return, terminates the hook.
  */
-extern int64_t accept              (uint32_t read_ptr,  uint32_t read_len,   uint32_t error_code);
+extern int64_t accept              (uint32_t read_ptr,  uint32_t read_len,   int64_t error_code);
 
 /**
  * Rollback the originating transaction, discard all hook state changes and emitted transactions.
@@ -37,7 +37,7 @@ extern int64_t accept              (uint32_t read_ptr,  uint32_t read_len,   uin
  * @param read_len The length of the string. May be 0.
  * @return Will never return, terminates the hook.
  */
-extern int64_t rollback            (uint32_t read_ptr,  uint32_t read_len,   int32_t error_code);
+extern int64_t rollback            (uint32_t read_ptr,  uint32_t read_len,   int64_t error_code);
 
 /**
  * Read a 20 byte account-id from the memory pointed to by read_ptr of length read_len and encode it to a base58-check
